Move PropGridPanel::Modify functions into their own file
